pyproject.toml
[project]
name = "crispr-project1"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dagster",
    "dagster-webserver",
    "dagster-pyspark",   # Connects Dagster orchestration to Spark execution
    "pyspark==3.5.0",
    "delta-spark==3.0.0",
    "duckdb",            # Keep for quick local lookups
    "psycopg2-binary",   # Keep for Postgres connectivity
    "python-dotenv>=1.2.1",
]
